💩 (lab3): Couldn't take it anymore and crushed rcore with one punch!

+++ os/src/task/priority.rs
@@ -1,6 +1,6 @@
 //! Priority definition
 
-type PriorityInner = u32;
+type PriorityInner = isize;
 
 /// Task Priority
 pub type Priority = PriorityImpl<PriorityInner>;
